refactor(frontend): renamed components files
